Remove Paragraph 3 custom section anchor link, the default is adequate.
Add one for the main document.

```
[[short-id]]
== Paragraph 3
```

This makes the § hover-link to paragraph 3 https://docs.gbif-uat.org/gbif-mou/en/#short-id

Without `[[short-id]]`, the default of #paragraph-3 will be used instead.

The intent is to provide a shorter/better link where that's useful, so I have added:

```
[[mou]]
== MEMORANDUM OF UNDERSTANDING FOR THE GLOBAL BIODIVERSITY INFORMATION FACILITY
```

We can then link to it with https://docs.gbif-uat.org/gbif-mou/en/#mou